Looking very good. I spent 2 hours trying to optimize the unsorted array setup only to realize it was not needed. The paper doesn't count it so I wont either. With that in mind im getting >4 speed up most of the time

+++ PSRS.h
@@ -19,12 +19,21 @@ pthread_barrier_t bar4;
 
 pthread_mutex_t print_mutex;
 
+
+
+
+
+
+
+
 struct phaseOneStruct {
     int* block; // sub array to be quick sorted
-    size_t size;
+    int size;
     int w; 
     int p; 
     int pSq;
+
+
 };
 
 struct phaseThreeFourStruct {
